Bug 26013 | Fix clang compilation warnings
[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/staging/PostPrimary/pbbam/": change = 163679]

+++ src/BaiIndexedBamReader.cpp
@@ -86,7 +86,7 @@ BaiIndexedBamReader& BaiIndexedBamReader::Interval(const GenomicInterval& interv
     auto header = bamFile_.Header();
     if (header.HasSequence(interval.Name())) {
         auto id = header.SequenceId(interval.Name());
-        if (id >= 0 && id < header.NumSequences()) {
+        if (id >= 0 && static_cast<size_t>(id) < header.NumSequences()) {
             htsIterator_.reset(bam_itr_queryi(htsIndex_.get(),
                                               id,
                                               interval.Start(),
